Start refactoring conversation layout a bit
