+++ assets/src/components/conversations/ConversationsContainer.tsx
@@ -1,55 +1,12 @@
 import React from 'react';
-import {Link} from 'react-router-dom';
-import {Box, Flex} from 'theme-ui';
-import {
-  Button,
-  colors,
-  Content,
-  Layout,
-  notification,
-  Result,
-  Sider,
-  Text,
-  Title,
-} from '../common';
-import {SmileOutlined} from '../icons';
+import {Box} from 'theme-ui';
+import {colors, Layout, notification, Sider, Text, Title} from '../common';
 import {sleep} from '../../utils';
-import Spinner from '../Spinner';
-import ChatMessage from './ChatMessage';
 import ConversationHeader from './ConversationHeader';
 import ConversationItem from './ConversationItem';
 import ConversationClosing from './ConversationClosing';
+import ConversationMessages from './ConversationMessages';
 import ConversationFooter from './ConversationFooter';
-
-const EmptyMessagesPlaceholder = () => {
-  return (
-    <Box my={4}>
-      <Result
-        status="success"
-        title="No messages"
-        subTitle="Nothing to show here! Take a well-earned break 😊"
-      />
-    </Box>
-  );
-};
-
-const GettingStartedRedirect = () => {
-  return (
-    <Box my={4}>
-      <Result
-        icon={<SmileOutlined />}
-        title="No messages"
-        subTitle="It looks like your widget hasn't been set up yet!"
-        extra={
-          <Link to="/account/getting-started">
-            <Button type="primary">Get Started</Button>
-          </Link>
-        }
-      />
-      ,
-    </Box>
-  );
-};
 
 type Props = {
   title?: string;
@@ -346,7 +303,8 @@ class ConversationsContainer extends React.Component<Props, State> {
 
     const loading = this.props.loading || this.state.loading;
     const isClosingSelected =
-      selectedConversationId && closing.indexOf(selectedConversationId) !== -1;
+      !!selectedConversationId &&
+      closing.indexOf(selectedConversationId) !== -1;
 
     return (
       <Layout style={{background: colors.white}}>
@@ -423,56 +381,15 @@ class ConversationsContainer extends React.Component<Props, State> {
             onDeleteConversation={this.handleDeleteConversation}
           />
 
-          <Content
-            style={{overflowY: 'scroll', opacity: isClosingSelected ? 0.6 : 1}}
-          >
-            {loading ? (
-              <Flex
-                sx={{
-                  flex: 1,
-                  justifyContent: 'center',
-                  alignItems: 'center',
-                  height: '100%',
-                }}
-              >
-                <Spinner size={40} />
-              </Flex>
-            ) : (
-              <Box
-                p={4}
-                backgroundColor={colors.white}
-                sx={{minHeight: '100%'}}
-              >
-                {messages.length ? (
-                  messages.map((msg: any, key: number) => {
-                    // Slight hack
-                    const next = messages[key + 1];
-                    const isMe = msg.user_id && msg.user_id === currentUser.id;
-                    const isLastInGroup = next
-                      ? msg.customer_id !== next.customer_id
-                      : true;
-
-                    // TODO: fix `isMe` logic for multiple agents
-                    return (
-                      <ChatMessage
-                        key={key}
-                        message={msg}
-                        customer={selectedCustomer}
-                        isMe={isMe}
-                        isLastInGroup={isLastInGroup}
-                        shouldDisplayTimestamp={isLastInGroup}
-                      />
-                    );
-                  })
-                ) : showGetStarted ? (
-                  <GettingStartedRedirect />
-                ) : (
-                  <EmptyMessagesPlaceholder />
-                )}
-                <div ref={(el) => (this.scrollToEl = el)} />
-              </Box>
-            )}
-          </Content>
+          <ConversationMessages
+            messages={messages}
+            currentUser={currentUser}
+            customer={selectedCustomer}
+            loading={loading}
+            isClosing={isClosingSelected}
+            showGetStarted={showGetStarted}
+            setScrollRef={(el) => (this.scrollToEl = el)}
+          />
 
           {selectedConversation && (
             // NB: the `key` forces a rerender so the input can clear
@@ -482,6 +399,8 @@ class ConversationsContainer extends React.Component<Props, State> {
               onSendMessage={this.handleSendMessage}
             />
           )}
+
+          {/* <Box sx={{width: 240, border: '1px solid black'}}>Test?</Box> */}
         </Layout>
       </Layout>
     );
